+++ ChangeLog.md
@@ -38,6 +38,7 @@ handling for several very different languages (proofs-of-concept).
  - [x] Clarify error message for missing concrete signature
  - [x] Fix exponential time bug in parser
  - [x] Allow prime characters in names after first position
+ - [x] Allow '.' to separate namespaces in imports/modules
 
 0.36.0 [2022.02.17]
 -------------------